backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
GDAL==3.9.0
ldap3==2.9.1
psycopg==3.2.9
pyasn1==0.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
